Database experiment for varying redundancy -- UNFINISHED.
Last experiments for varying MFT (ran in jupiterace server)
showed that standard MC still outperforms RESTART for the rarest event
(MFT ~ 128000), even with the new exponential thresholds selection.

On the other hand, fast tests in laptop suggested varying the system
redundancy is the way to go, just like V-A suggested.
Currently automatizing the composition functions construction used by
the --acomp strategy, for varying redundancy values.

Next immediate step is to implement the constructor of the "natural
function", i.e. the one that could be automatically derived from the
property specification. That needs combinations :(
Check "comp_fun_fine()" defined in experiments/rate_red/ifuns.sh

+++ include/core_typedefs.h
@@ -142,7 +142,7 @@ typedef  std::bitset<MAX_NUM_CLOCKS>                                 Bitflag;
 /// Primitive type used to assess the importance of a single *concrete* state
 /// @warning This bounds the number of representable importance levels
 ///
-typedef  unsigned int                                        ImportanceValue;
+typedef  size_t                                              ImportanceValue;
 
 /// Vector of \ref ImportanceValue "importance values"
 typedef  std::vector< ImportanceValue >                        ImportanceVec;
